requirements.txt
Django==1.11.3
djangorestframework==3.6.3
pytz==2017.2
six==1.10.0
